+++ src/Ext/BuildingType/Body.cpp
@@ -264,6 +264,8 @@ void BuildingTypeExt::LoadFromINIFile(CCINIClass* const pINI)
 			this->PowersUp_Buildings.emplace_back(pPowerUpType);
 	}
 
+	this->SetTabBySelecting.Read(exINI, pSection, "SetTabBySelecting");
+
 	if (pThis->NumberOfDocks > 0)
 	{
 		std::optional<DirType> empty;
@@ -433,6 +435,7 @@ void BuildingTypeExt::Serialize(T& Stm)
 		.Process(this->TurretAnim_FiringFrames)
 		.Process(this->StartFacing)
 		.Process(this->StartFacing_Random)
+		.Process(this->SetTabBySelecting)
 
 		// Ares 0.2
 		.Process(this->CloningFacility)